# Conflicts:
#	ruoyi-common/ruoyi-common-core/src/main/java/com/ruoyi/common/core/domain/http/Channel.java
#	ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/controller/CustomerController.java
#	ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/service/IMerchantService.java
#	ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/service/impl/MerchantServiceImpl.java

+++ ruoyi-modules/ruoyi-system/src/main/java/com/ruoyi/system/service/IChannelService.java
@@ -18,7 +18,7 @@ public interface IChannelService
      * @param id 渠道配置主键
      * @return 渠道配置
      */
-    public Channel selectChannelById(Long id);
+    public Channel selectChannelById(Integer id);
 
     /**
      * 查询渠道配置列表
@@ -34,7 +34,7 @@ public interface IChannelService
      * @param channel 渠道配置
      * @return 结果
      */
-    public Long insertChannel(Channel channel);
+    public int insertChannel(Channel channel);
 
     /**
      * 修改渠道配置
